Improvements to fetch, events, and url match
Fetching paginated results no longer makes an extra call for the final page, instead determining the next page url from response headers. This also makes the urls in the fetch requests more sensible, since they don't have to include extra parameters so paginated fetch can add `&page=1` etc.

The time and date string constructors have been added to the Event object, which is now a class.

The script now checks the current page url for the `/courses/sis_course_id:123456` format. If it encounters that format, it redirects to the standard format and prompts the user to run the bookmarklet again. It also alerts the user if the script was unable to get any ID from the url.

Finally, the script now prompts the user to decide if they want to include descriptions for assignments and calendar events.